Able to register user
