+++ app/src/main/java/com/example/caterpillar/LoginActivity.java
@@ -7,27 +7,21 @@ import android.view.View;
 import android.content.Intent;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
+
 import android.widget.Toast;
 
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.json.JSONArray;
+
 
 import com.github.nkzawa.emitter.Emitter;
-import com.github.nkzawa.engineio.client.Transport;
-import com.github.nkzawa.socketio.client.Manager;
 import com.github.nkzawa.socketio.client.Socket;
 import com.server.interaction.SocketManager;
 
-import java.util.HashMap;
-import java.util.Map;
 
 public class LoginActivity extends AppCompatActivity {
     private EditText mUsernameView;
     private EditText mPasswordView;
-    private String mUsername;
-
     private Socket mSocket;
 
     @Override
@@ -36,21 +30,7 @@ public class LoginActivity extends AppCompatActivity {
         setContentView(R.layout.activity_login);
         SocketManager app = (SocketManager) getApplication();
         mSocket = app.getmSocket();
-        mSocket.io().on(Manager.EVENT_TRANSPORT, new Emitter.Listener() {
-            @Override
-            public void call(Object... args) {
-                Transport transport = (Transport) args[0];
-                transport.on(Transport.EVENT_ERROR, new Emitter.Listener() {
-                    @Override
-                    public void call(Object... args) {
-                        Exception e = (Exception) args[0];
-                        Log.e("socketConnection", "Transport error " + e);
-                        e.printStackTrace();
-                        e.getCause().printStackTrace();
-                    }
-                });
-            }
-        });
+
         // Set up the login details
         mUsernameView = findViewById(R.id.input_user);
         mPasswordView = findViewById(R.id.input_password);
