+++ app/src/main/res/layout/activity_sign_up.xml
@@ -49,7 +49,7 @@
         android:layout_marginRight="480dp"
         android:layout_marginBottom="58dp"
         android:textColorHint="@color/colorPrimary"
-        app:layout_constraintBottom_toTopOf="@+id/login"
+        app:layout_constraintBottom_toTopOf="@+id/register_button"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textInputLayout3">
